+++ tests/test_scorers.py
@@ -2,7 +2,7 @@ import pytest
 from spacy.tokens import Span
 
 import edsnlp
-from edsnlp.scorers.ner import NerExactScorer, NerTokenScorer
+from edsnlp.scorers.ner import NerExactScorer, NerOverlapScorer, NerTokenScorer
 from edsnlp.scorers.span_attributes import SpanAttributeScorer
 
 
@@ -11,17 +11,29 @@ def gold_and_pred():
     nlp = edsnlp.blank("eds")
 
     gold_doc1 = nlp.make_doc("Le patient a le covid 19.")
-    gold_doc1.ents = [Span(gold_doc1, 4, 6, label="covid")]
-    gold_doc2 = nlp.make_doc("Corona: positif. Le cvid est une maladie.")
+    gold_doc1.ents = [
+        Span(gold_doc1, 4, 6, label="covid"),  # le covid
+    ]
+    pred_doc1 = nlp.make_doc("Le patient a le covid 19.")
+    pred_doc1.ents = [
+        Span(pred_doc1, 4, 6, label="covid"),  # le covid
+    ]
+    gold_doc2 = nlp.make_doc(
+        "Corona: positif. Le cvid est une maladie très très grave."
+    )
     gold_doc2.ents = [
-        Span(gold_doc2, 0, 1, label="covid"),
-        Span(gold_doc2, 5, 6, label="covid"),
+        Span(gold_doc2, 0, 1, label="covid"),  # Corona
+        Span(gold_doc2, 5, 6, label="covid"),  # cvid
+        Span(gold_doc2, 8, 12, label="disease"),  # maladie très très grave
     ]
 
-    pred_doc1 = nlp.make_doc("Le patient a le covid 19.")
-    pred_doc1.ents = [Span(pred_doc1, 4, 6, label="covid")]
-    pred_doc2 = nlp.make_doc("Corona: positif. Le cvid est une maladie.")
-    pred_doc2.ents = [Span(pred_doc2, 0, 2, label="covid")]
+    pred_doc2 = nlp.make_doc(
+        "Corona: positif. Le cvid est une maladie très très grave."
+    )
+    pred_doc2.ents = [
+        Span(pred_doc2, 0, 2, label="covid"),  # Corona:
+        Span(pred_doc2, 8, 9, label="disease"),  # maladie
+    ]
 
     return [gold_doc1, gold_doc2], [pred_doc1, pred_doc2]
 
@@ -30,11 +42,11 @@ def test_exact_ner_scorer(gold_and_pred):
     scorer = NerExactScorer("ents", filter_expr="'vid' in doc.text")
     ner_exact_score = scorer(*gold_and_pred)
     assert ner_exact_score["micro"] == {
-        "p": 0.5,
-        "r": 1 / 3,
-        "f": 0.4,
-        "support": 3,
-        "positives": 2,
+        "f": 0.2857142857142857,
+        "p": 0.3333333333333333,
+        "positives": 3,
+        "r": 0.25,
+        "support": 4,
         "tp": 1,
     }
 
@@ -43,12 +55,70 @@ def test_token_ner_scorer(gold_and_pred):
     scorer = NerTokenScorer("ents", filter_expr="'vid' in doc.text")
     ner_exact_score = scorer(*gold_and_pred)
     assert ner_exact_score["micro"] == {
-        "f": 0.75,
-        "p": 0.75,
+        "f": 0.6153846153846154,
+        "p": 0.8,
+        "positives": 5,
+        "r": 0.5,
+        "support": 8,
+        "tp": 4,
+    }
+
+
+def test_overlap_ner_scorer_any(gold_and_pred):
+    scorer = NerOverlapScorer(
+        "ents", threshold=0.00001, filter_expr="'vid' in doc.text"
+    )
+    # pred entities: [le covid, Corona:, maladie] => 3
+    # gold entities: [le covid, Corona, cvid, maladie très très grave] => 4
+    # tp: [le covid, Corona, maladie] => 3
+    ner_exact_score = scorer(*gold_and_pred)
+    assert ner_exact_score["micro"] == {
+        "f": 0.8571428571428572,
+        "p": 1,
+        "positives": 3,
         "r": 0.75,
         "support": 4,
-        "positives": 4,
         "tp": 3,
+    }
+
+
+def test_overlap_ner_scorer_half(gold_and_pred):
+    scorer = NerOverlapScorer(
+        "ents",
+        threshold=0.5,
+        filter_expr="'vid' in doc.text",
+    )
+    # pred entities: [le covid, Corona:, maladie] => 3
+    # gold entities: [le covid, Corona, cvid, maladie très très grave] => 4
+    # tp: [le covid, Corona] => 2
+    ner_exact_score = scorer(*gold_and_pred)
+    assert ner_exact_score["micro"] == {
+        "f": 0.5714285714285714,
+        "p": 0.6666666666666666,
+        "positives": 3,
+        "r": 0.5,
+        "support": 4,
+        "tp": 2,
+    }
+
+
+def test_overlap_ner_scorer_full(gold_and_pred):
+    scorer = NerOverlapScorer(
+        "ents",
+        threshold=1.0,
+        filter_expr="'vid' in doc.text",
+    )
+    ner_exact_score = scorer(*gold_and_pred)
+    # pred entities: [le covid, Corona:, maladie] => 3
+    # gold entities: [le covid, Corona, cvid, maladie très très grave] => 4
+    # tp: [le covid] => 2
+    assert ner_exact_score["micro"] == {
+        "f": 0.2857142857142857,
+        "p": 0.3333333333333333,
+        "positives": 3,
+        "r": 0.25,
+        "support": 4,
+        "tp": 1,
     }